% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.label.R
\name{bicluster.label}
\alias{bicluster.label}
\title{generate cluster label for bicluster}
\usage{
bicluster.label(X)
}
\arguments{
\item{X}{The data matrix to be clustered, containing rownames and colnames.}
}
\value{
bicluster labels
}
\description{
generate cluster label for bicluster
}
\examples{

x <-matrix(c(1,2,3,4),2,2)
rownames(x) <-c('row1','row2')
colnames(x) <-c('col1','col2')
bicluster.label(x)
}
