% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bicluster.label.R
\name{cluster.label}
\alias{cluster.label}
\title{generate cluster label for cluster using row or col names}
\usage{
cluster.label(X, labels = "col")
}
\arguments{
\item{X}{The data matrix to be clustered}

\item{labels}{using row or col names}
}
\value{
cluster labels
}
\description{
generate cluster label for cluster using row or col names
}
\examples{
x.rowlab<-cluster.label(x,labels = 'row')$num
x.collab<-cluster.label(x)$num
# adjustedRandIndex(x.rowlab,x.collab)

}
