% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_assignment.R
\name{cluster_assign}
\alias{cluster_assign}
\title{\code{cluster_assign}}
\usage{
cluster_assign(X, k, result, phi = 0.5)
}
\arguments{
\item{X}{Data matrix}

\item{k}{the number of nearest-neighbors}

\item{result}{the list, output of bi_ADMM function.}
}
\description{
\code{cluster_assign}
}
