% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen.R
\name{data_gen}
\alias{data_gen}
\title{Simulated data set generation}
\usage{
data_gen(
  seed.cluster = 123,
  seed.data = 654,
  n,
  true_p,
  p,
  mu.lower = -10,
  mu.upper = 10,
  theta = 2,
  theta_noise = 0.1,
  row_group = 4,
  col_group = 4
)
}
\arguments{
\item{seed.cluster}{seed.cluster is used to control the cluster assignment}

\item{seed.data}{seed.data is used to control the data generation given clustering structure}

\item{n}{number of subjects}

\item{true_p}{number of informative features}

\item{p}{number of total features}

\item{mu.lower}{lower bound on the mean of the normal distribution}

\item{mu.upper}{upper bound on the mean of the normal distribution}

\item{theta}{the standard deviation of the x, which is sampled from a normal distribution}

\item{theta_noise}{the standard deviation of noisy features}

\item{row_group}{number of row clusters}

\item{col_group}{number of informative column clusters}
}
\value{
Output is the simulated dataset
}
\description{
\code{data_gen} generates dataset contains uninformative features in the simulation study.
}
\examples{
set.seed(123)
X = data_gen(n = 100, true_p=40, p = 80)
}
