% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_selection_diagnostic.R
\name{feature.selection.diagnostic}
\alias{feature.selection.diagnostic}
\title{Evaluate diagnostic statistics for estimated clusters}
\usage{
feature.selection.diagnostic(X, A, tol = 1e-06)
}
\arguments{
\item{X}{Data matrix}

\item{A}{Biclustering result}

\item{tol}{threshold of noisy variable}
}
\value{
diagnostic statistics
}
\description{
Evaluate diagnostic statistics for estimated clusters
}
\examples{
set.seed(123)
X = data_gen(n = 100, true_p=40, p = 80)
#Pretend A is the Biclustering result
set.seed(111)
A = data_gen(n = 100, true_p=40, p = 80)
feature.selection.diagnostic(X, res1$A)
}
