% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_validation_biclustering.R
\name{predict_bi_validate}
\alias{predict_bi_validate}
\title{Predict new data for biclustering problem}
\usage{
predict_bi_validate(
  X,
  clusters_row,
  clusters_col,
  X_validate,
  label_valdiate = NULL,
  label.known = 1
)
}
\arguments{
\item{X}{(n by p)}

\item{clusters_row}{cluster assignment for each subject}

\item{clusters_col}{cluster assignment for each feature}

\item{data_validate}{validate dataset}
}
\description{
report Rand index and variable stability in validate dataset
}
