% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biADMM.R
\name{proj_Linf_m}
\alias{proj_Linf_m}
\title{Projection of infinity norm}
\usage{
proj_Linf_m(x, tau)
}
\arguments{
\item{x}{data matrix}

\item{tau}{tau}
}
\value{
px
}
\description{
Projection of infinity norm
}
