% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biADMM.R
\name{sign.c}
\alias{sign.c}
\title{Adjusts the sign of a vector according to specified conditions.}
\usage{
\method{sign}{c}(px, x)
}
\arguments{
\item{px}{The condition for adjusting the sign.}

\item{x}{The vector whose sign needs adjustment.}
}
\value{
a
}
\description{
Adjusts the sign of a vector according to specified conditions.
}
