% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_assignment.R
\name{tri2vec}
\alias{tri2vec}
\title{Cluster_assignment: get the cluster labels after ADMM convex-biclustering algorithm}
\usage{
tri2vec(i, j, n)
}
\description{
The function: tri2vec, vec2tri, kernel_weights, knn_weights, get_subgroup_means_full, biclust_smooth,
create_adjacency, find_clusters are from the cvxbiclustr package.
}
